requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
django-auto-logout==0.5.1
django-axes==6.4.0
django-crispy-forms==2.1
django-environ==0.11.2
django-formtools==2.5.1
django-otp==1.3.0
django-phonenumber-field==7.3.0
django-recaptcha==4.0.0
django-two-factor-auth==1.16.0
django-utils-six==2.0
idna==3.6
oauthlib==3.2.2
phonenumberslite==8.13.31
pycparser==2.21
PyJWT==2.8.0
pypng==0.20220715.0
python3-openid==3.2.0
pytz==2024.1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
